requirements.txt
Flask==0.12.2
line-bot-sdk
beautifulsoup4